requirements.txt
Flask==3.0.3
openai==1.30.1
python-dotenv==1.0.0
httpx==0.27.0
gunicorn==21.2.0 